Refactor CrossChainStore
